Version 1.0
Was unable to commit progress due to being on vacation while coding all
this. This is the first fully complete version of the classifier. It
currently has no support for discrete data and treats all data as
continuous but should soon. A brief changelog is as follows:

Separates classifiers from data which are defined in the second sheet of
the data
Data now is expected to have a second sheet containing some meta data as
follows: first row is expected to have attribute names, second row is to
have data type (unused), third row is to have 'classifier' in whatever
cell is the classifier (Any value will do but best to say something
specific to users anyways)
Now generates training data on one of 3 different ways: first n amount,
random n amount, or every total/nth value for n amounts.
Now counts instances of occurrences of different attribute values and
stores them
Now computes value of all possible classifiers and decides the most
likely and stores in guessedClassifiers
Now outputs an excel file with the guessed classifications in the first
sheet and the second contains the trainingData
Has several print functions for debugging purposes

A current todo list in order of importance which may not all be done:

Create a GUI frontend (Simple would just be used to select the file,
changes values in second sheet of excel file, choose a secondary
location for output file to be copied to so it is not just in the
results folder and can be found easily by user)
Compute the frequency values in generateClassifier instead of classify
so they all can be done in one pass and then also store possible
classification frequencies in a global location which will make
continuous data support easier to implement
Make a classifier generator for continuous data (gaussian naive bayes)
Make generateClassifier a class that decides which classifier to call
for each attribute based on the data provided in sheet 2 row 2
Fix cellStyle coloring for excel files
Cleanup error logs
Privatize data and have gets/make entire project more abstract and
OO/more import and implements friendly for future use.
Bundle powerpoint going to be used for class for explanation of Naive
Bayes and program
Bundle user and technical manuals
Document code (This is not important, right?)