Fix bugfix and possible cheat
Now calculates total possible outcomes when generating training data so
it does not cheat and use all possible outcomes from the data when
imported.
Moved output over so it does not overwrite attribute occurrence name.

Also forgot a todo I want to do which is to a a generate training data
that generates it from an external file.