New sheet in excel file
Now has a third sheet that contains a nice readable version of the
likelihood of any value an attribute had for all possible classifiers
and then the likelihood of each classifier.
Also fixed a few times where the word frequency was used after the
frequencies have been turned into likelihoods.